requirements.txt
# Core AI Dependencies
google-adk>=0.3.0
google-genai>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Testing Framework
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Frontend
streamlit>=1.30.0
